requirements.txt
asgiref==3.7.2
Choco==1.0.5
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
djangorestframework==3.14.0
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
